+++ app/scripts/controllers/comparison_controller.js
@@ -2,6 +2,7 @@
 import _ from "lodash"
 import settings from "@/settings"
 import { stringifyFunc } from "@/stringify.js"
+import { translateAttribute } from "@/util"
 
 const korpApp = angular.module("korpApp")
 
@@ -42,7 +43,7 @@ korpApp.directive("compareCtrl", () => ({
                             stringify = stringifyFunc(reduceAttrName)
                         } else if (attributes[reduceAttrName].translation) {
                             stringify = (value) =>
-                                util.translateAttribute($rootScope.lang, attributes[reduceAttrName].translation, value)
+                                translateAttribute($rootScope.lang, attributes[reduceAttrName].translation, value)
                         }
                     }
                     s.stringify = [stringify]