+++ app/scripts/filter_directives.js
@@ -5,6 +5,7 @@
  */
 import _ from "lodash"
 import settings from "@/settings"
+import { regescape, translateAttribute } from "@/util"
 
 const korpApp = angular.module("korpApp")
 
@@ -115,9 +116,8 @@ korpApp.directive("globalFilter", [
 
             scope.isSelectedList = (value) => scope.selected.includes(value)
 
-            scope.translateAttribute = (value) => {
-                return util.translateAttribute(scope.lang, scope.attrDef.settings.translation, value)
-            }
+            scope.translateAttribute = (value) =>
+                translateAttribute(scope.lang, scope.attrDef.settings.translation, value)
         },
     }),
 ])