+++ CHANGELOG.md
@@ -7,7 +7,15 @@
 - Extended support for description on frontpage, with corpus updates [#341](https://github.com/spraakbanken/korp-frontend/issues/341)
 - The "medial part" option now includes first/last parts also for lemgram search [#347](https://github.com/spraakbanken/korp-frontend/issues/347)
 - Improved UI reactivity for Simple search
-- Removed globals `settings`, `currentMode`, `currentModeParallel`, `model`, `CorpusListing`, `ParallelCorpusListing` – import them instead
+- Removed globals – import them instead:
+  - `settings`
+  - `currentMode`
+  - `currentModeParallel`
+  - `model`
+  - `CorpusListing`
+  - `ParallelCorpusListing`
+  - `regescape`
+  - `unregescape`
 - Renamed localization functions (just like the template filters):
   - `getLocaleString` to `loc`
   - `getLocaleStringObject` to `locObj`
@@ -20,6 +28,8 @@
 - Converted from `window.util` props to exported members:
   - `loc`, `locObj`
   - lemgram/saldo functions
+  - `setDownloadLinks`
+  - `translateAttribute`
 
 ### Fixed